Don't URI.encode URLs that don't need it
This prevents URLs that are already URLEncoded from getting
doubly-encoded, breaking the resulting shortened URL. This is the case
most of the time with URLs copied from popular browsers' address
bar, at least with non-latin1 characters. Tested with tinyurl,
untested with the other services, but behaviour should be the same.